0.0.2 update

+++ include/mtwister.h
@@ -1,16 +1,18 @@
 #ifndef __MTWISTER_H
 #define __MTWISTER_H
 
+#include <inttypes.h>
+
 #define STATE_VECTOR_LENGTH 624
 #define STATE_VECTOR_M      397 /* changes to STATE_VECTOR_LENGTH also require changes to this */
 
 typedef struct tagMTRand {
-  unsigned long mt[STATE_VECTOR_LENGTH];
+  uint_fast32_t mt[STATE_VECTOR_LENGTH];
   int index;
 } MTRand;
 
-MTRand seedRand(unsigned long seed);
-unsigned long genRandLong(MTRand* rand);
+MTRand seedRand(uint_fast32_t seed);
+uint_fast32_t genRandLong(MTRand* rand);
 double genRand(MTRand* rand);
 
 #define UPPER_MASK		0x80000000
@@ -18,7 +20,7 @@ double genRand(MTRand* rand);
 #define TEMPERING_MASK_B	0x9d2c5680
 #define TEMPERING_MASK_C	0xefc60000
 
-inline static void m_seedRand(MTRand* rand, unsigned long seed) {
+inline static void m_seedRand(MTRand* rand, uint_fast32_t seed) {
   /* set initial seeds to mt[STATE_VECTOR_LENGTH] using the generator
    * from Line 25 of Table 1 in: Donald Knuth, "The Art of Computer
    * Programming," Vol. 2 (2nd Ed.) pp.102.
@@ -32,7 +34,7 @@ inline static void m_seedRand(MTRand* rand, unsigned long seed) {
 /**
 * Creates a new random number generator from a given seed.
 */
-MTRand seedRand(unsigned long seed) {
+MTRand seedRand(uint_fast32_t seed) {
   MTRand rand;
   m_seedRand(&rand, seed);
   return rand;
@@ -41,10 +43,10 @@ MTRand seedRand(unsigned long seed) {
 /**
  * Generates a pseudo-randomly generated long.
  */
-unsigned long genRandLong(MTRand* rand) {
+uint_fast32_t genRandLong(MTRand* rand) {
 
-  unsigned long y;
-  static unsigned long mag[2] = {0x0, 0x9908b0df}; /* mag[x] = x * 0x9908b0df for x = 0,1 */
+  uint_fast32_t y;
+  static uint_fast32_t mag[2] = {0x0, 0x9908b0df}; /* mag[x] = x * 0x9908b0df for x = 0,1 */
   if(rand->index >= STATE_VECTOR_LENGTH || rand->index < 0) {
     /* generate STATE_VECTOR_LENGTH words at a time */
     int kk;
@@ -75,7 +77,7 @@ unsigned long genRandLong(MTRand* rand) {
  * Generates a pseudo-randomly generated double in the range [0..1].
  */
 double genRand(MTRand* rand) {
-  return((double)genRandLong(rand) / (unsigned long)0xffffffff);
+  return((double)genRandLong(rand) / (uint_fast32_t)0xffffffff);
 }
 
 #endif /* #ifndef __MTWISTER_H */
